+++ benchmarks/AutoClusterBenchmark/Configuration.cs
@@ -37,7 +37,7 @@ public static class Configuration
         var helloProps = Props.FromProducer(() => new WorkerActor());
         return ClusterConfig
             .Setup("mycluster", clusterProvider, identityLookup)
-            .WithClusterContextProducer(cluster => new ExperimentalClusterContext(cluster))
+            .WithClusterContextProducer(cluster => new DefaultClusterContext(cluster))
             .WithClusterKind("hello", helloProps)
             .WithGossipFanOut(3);
     }
@@ -119,6 +119,7 @@ public static class Configuration
     private static ActorSystemConfig GetMemberActorSystemConfig()
     {
         var config = new ActorSystemConfig()
+            .WithSharedFutures()
             .WithDeadLetterThrottleCount(3)
             .WithDeadLetterThrottleInterval(TimeSpan.FromSeconds(1))
             .WithDeadLetterRequestLogging(false);
